simple seuare

+++ lib/providers/authentication_provider.dart
@@ -1,4 +1,5 @@
 //Packages
+//import 'package:chatify_app/models/chat_user.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:get_it/get_it.dart';
@@ -25,22 +26,22 @@ class AuthenticationProvider extends ChangeNotifier {
     _auth.authStateChanges().listen((_user) {
       if (_user != null) {
         _databaseService.updateUserLastSeenTime(_user.uid);
-        _databaseService.getUser(_user.uid).then((_snapshort) {
+        _databaseService.getUser(_user.uid).then((_snapshot) {
           Map<String, dynamic> _userData =
-              _snapshort.data()! as Map<String, dynamic>;
-          user = ChatUser.fromJSON(
-            {
-              "uid": _user.uid,
-              "name": _userData["name"],
-              "email": _userData["email"],
-              "last_active": _userData["last_active"],
-              "image": _userData["image"],
-            },
-          );
+              _snapshot.data()! as Map<String, dynamic>;
+          user = ChatUser.fromJSON({
+            "uid": _user.uid,
+            "name": _userData["name"],
+            "email": _userData["email"],
+            "last_active": _userData["last_active"],
+            "image": _userData["image"],
+          });
           _navigationService.removeAndNavigateToRoute('/home');
         });
       } else {
-        _navigationService.removeAndNavigateToRoute('/login');
+        if (_navigationService.getCurrentRoute() != '/login') {
+          _navigationService.removeAndNavigateToRoute('/login');
+        }
       }
     });
   }
@@ -54,11 +55,62 @@ class AuthenticationProvider extends ChangeNotifier {
         email: _email,
         password: _password,
       );
-      print(_auth.currentUser);
     } on FirebaseAuthException {
       print("Error logging user into Firebase");
     } catch (e) {
       print(e);
     }
   }
+
+  Future<String?> registerUserUsingEmailAndPassword(
+    String _email,
+    String _password,
+  ) async {
+    try {
+      UserCredential _credentials = await _auth.createUserWithEmailAndPassword(
+        email: _email,
+        password: _password,
+      );
+      return _credentials.user!.uid;
+    } on FirebaseAuthException {
+      print("Error registering user.");
+    } catch (e) {
+      print(e);
+    }
+  }
+
+  
+
+  Future<void> logout() async {
+    try {
+      await _auth.signOut();
+    } catch (e) {
+      print(e);
+    }
+  }
+
+  // Future<String?> registerUserUsingEmailAndPassword(
+  //   String _email,
+  //   String _password,
+  // ) async {
+  //   try {
+  //     UserCredential _credentials = await _auth.createUserWithEmailAndPassword(
+  //       email: _email,
+  //       password: _password,
+  //     );
+  //     return _credentials.user!.uid;
+  //   } on FirebaseAuthException {
+  //     print("Error registering user.");
+  //   } catch (e) {
+  //     print(e);
+  //   }
+  // }
+
+  // Future<void> logout() async {
+  //   try {
+  //     await _auth.signOut();
+  //   } catch (e) {
+  //     print(e);
+  //   }
+  // }
 }
